add a negative cache to regexp decoder
Often, regexp decoder can run repeatedly on the same input and skip them
with regexp filter. An common example is matching cgroup path in a chain like so:

```
- name: cgroup
- name: regexp
  regexps:
  - ^.*(system.slice).*$
```

Anything that is not in system.slice will be skipped. In some cases
where we filter only a small subset of inputs, the overhead of regexp matching
can often be noticable.

We add a cache here to test for input that would produce ErrSkipLabelSet and
skip regex matching on them to reduce the work done on regexp matching.
The cache size is customizable with `skip_cache_size` configuration.

Signed-off-by: Daniel Dao <dqminh@cloudflare.com>