add a lru cache to regexp decoder
Often, regexp decoder can run repeatedly on the same input. An common
example is matching cgroup path in a chain like so:

```
- name: cgroup
- name: regexp
  regexps:
  - ^.*(system.slice).*$
```

In these cases, the overhead of regexp matching can often be noticable.
Since the input does not vary too often, we can add a cache here to skip
the overhead of regex matching. This is implemented as an LRU cache to
prevent the cache from growing infinitely, especially in cases where we
have many short-live cgroups. The previous example can be written aborts

```
- name: cgroup
- name: regexp
  lru_cache_size: 1024
  regexps:
  - ^.*(system.slice).*$
```

Signed-off-by: Daniel Dao <dqminh@cloudflare.com>